+++ packages/svelte/src/compiler/phases/3-transform/client/transform-client.js
@@ -84,7 +84,10 @@ export function client_component(source, analysis, options) {
 		},
 		legacy_reactive_statements: new Map(),
 		metadata: {
-			template_needs_import_node: false,
+			context: {
+				template_needs_import_node: false,
+				template_contains_script_tag: false
+			},
 			namespace: options.namespace,
 			bound_contenteditable: false
 		},
@@ -262,6 +265,7 @@ export function client_component(source, analysis, options) {
 		...legacy_reactive_declarations,
 		...group_binding_declarations,
 		.../** @type {import('estree').Statement[]} */ (instance.body),
+		analysis.runes ? b.empty : b.stmt(b.call('$.init')),
 		.../** @type {import('estree').Statement[]} */ (template.body),
 		...static_bindings
 	]);