+++ packages/svelte/src/compiler/phases/3-transform/client/types.d.ts
@@ -47,9 +47,22 @@ export interface ComponentClientTransformState extends ClientTransformState {
 	readonly template: string[];
 	readonly metadata: {
 		namespace: Namespace;
-		/** `true` if the HTML template needs to be instantiated with `importNode` */
-		template_needs_import_node: boolean;
 		bound_contenteditable: boolean;
+		/**
+		 * Stuff that is set within the children of one `create_block` that is relevant
+		 * to said `create_block`. Shouldn't be destructured or otherwise spread unless
+		 * inside `create_block` to keep the object reference intact (it's also nested
+		 * within `metadata` for this reason).
+		 */
+		context: {
+			/** `true` if the HTML template needs to be instantiated with `importNode` */
+			template_needs_import_node: boolean;
+			/**
+			 * `true` if HTML template contains a `<script>` tag. In this case we need to invoke a special
+			 * template instantiation function (see `create_fragment_with_script_from_html` for more info)
+			 */
+			template_contains_script_tag: boolean;
+		};
 	};
 	readonly preserve_whitespace: boolean;