+++ packages/svelte/src/internal/client/render.js
@@ -4,6 +4,7 @@ import {
 	child,
 	clone_node,
 	create_element,
+	empty,
 	init_operations,
 	map_get,
 	map_set,
@@ -23,14 +24,18 @@ import {
 	PassiveDelegatedEvents,
 	DelegatedEvents,
 	AttributeAliases,
-	namespace_svg,
-	namespace_html
+	namespace_svg
 } from '../../constants.js';
-import { create_fragment_from_html, insert, reconcile_html, remove } from './reconciler.js';
+import {
+	create_fragment_from_html,
+	create_fragment_with_script_from_html,
+	insert,
+	reconcile_html,
+	remove
+} from './reconciler.js';
 import {
 	render_effect,
 	destroy_signal,
-	get,
 	is_signal,
 	push_destroy_fn,
 	execute_effect,
@@ -71,24 +76,37 @@ const all_registerd_events = new Set();
 /** @type {Set<(events: Array<string>) => void>} */
 const root_event_handles = new Set();
 
-/** @returns {Text} */
-export function empty() {
-	return document.createTextNode('');
-}
-
 /**
  * @param {string} html
- * @param {boolean} is_fragment
+ * @param {boolean} return_fragment
  * @returns {() => Node}
  */
 /*#__NO_SIDE_EFFECTS__*/
-export function template(html, is_fragment) {
+export function template(html, return_fragment) {
 	/** @type {undefined | Node} */
 	let cached_content;
 	return () => {
 		if (cached_content === undefined) {
 			const content = create_fragment_from_html(html);
-			cached_content = is_fragment ? content : /** @type {Node} */ (child(content));
+			cached_content = return_fragment ? content : /** @type {Node} */ (child(content));
+		}
+		return cached_content;
+	};
+}
+
+/**
+ * @param {string} html
+ * @param {boolean} return_fragment
+ * @returns {() => Node}
+ */
+/*#__NO_SIDE_EFFECTS__*/
+export function template_with_script(html, return_fragment) {
+	/** @type {undefined | Node} */
+	let cached_content;
+	return () => {
+		if (cached_content === undefined) {
+			const content = create_fragment_with_script_from_html(html);
+			cached_content = return_fragment ? content : /** @type {Node} */ (child(content));
 		}
 		return cached_content;
 	};
@@ -96,17 +114,35 @@ export function template(html, is_fragment) {
 
 /**
  * @param {string} svg
- * @param {boolean} is_fragment
+ * @param {boolean} return_fragment
  * @returns {() => Node}
  */
 /*#__NO_SIDE_EFFECTS__*/
-export function svg_template(svg, is_fragment) {
+export function svg_template(svg, return_fragment) {
 	/** @type {undefined | Node} */
 	let cached_content;
 	return () => {
 		if (cached_content === undefined) {
 			const content = /** @type {Node} */ (child(create_fragment_from_html(`<svg>${svg}</svg>`)));
-			cached_content = is_fragment ? content : /** @type {Node} */ (child(content));
+			cached_content = return_fragment ? content : /** @type {Node} */ (child(content));
+		}
+		return cached_content;
+	};
+}
+
+/**
+ * @param {string} svg
+ * @param {boolean} return_fragment
+ * @returns {() => Node}
+ */
+/*#__NO_SIDE_EFFECTS__*/
+export function svg_template_with_script(svg, return_fragment) {
+	/** @type {undefined | Node} */
+	let cached_content;
+	return () => {
+		if (cached_content === undefined) {
+			const content = /** @type {Node} */ (child(create_fragment_from_html(`<svg>${svg}</svg>`)));
+			cached_content = return_fragment ? content : /** @type {Node} */ (child(content));
 		}
 		return cached_content;
 	};
@@ -172,11 +208,22 @@ const space_template = template(' ', false);
 const comment_template = template('<!>', true);
 
 /**
- * @param {null | Text | Comment | Element} anchor
+ * @param {Text | Comment | Element | null} anchor
  */
 /*#__NO_SIDE_EFFECTS__*/
 export function space(anchor) {
-	return open(anchor, true, space_template);
+	/** @type {Node | null} */
+	var node = /** @type {any} */ (open(anchor, true, space_template));
+	// if an {expression} is empty during SSR, there might be no
+	// text node to hydrate (or an anchor comment is falsely detected instead)
+	//  — we must therefore create one
+	if (current_hydration_fragment !== null && node?.nodeType !== 3) {
+		node = empty();
+		// @ts-ignore in this case the anchor should always be a comment,
+		// if not something more fundamental is wrong and throwing here is better to bail out early
+		anchor.parentElement.insertBefore(node, anchor);
+	}
+	return node;
 }
 
 /**
@@ -188,6 +235,8 @@ export function comment(anchor) {
 }
 
 /**
+ * Assign the created (or in hydration mode, traversed) dom elements to the current block
+ * and insert the elements into the dom (in client mode).
  * @param {Element | Text} dom
  * @param {boolean} is_fragment
  * @param {null | Text | Comment | Element} anchor
@@ -2826,7 +2875,9 @@ export function mount(component, options) {
 	const container = options.target;
 	const block = create_root_block(options.intro || false);
 	const first_child = /** @type {ChildNode} */ (container.firstChild);
-	const hydration_fragment = get_hydration_fragment(first_child);
+	// Call with insert_text == true to prevent empty {expressions} resulting in an empty
+	// fragment array, resulting in a hydration error down the line
+	const hydration_fragment = get_hydration_fragment(first_child, true);
 	const previous_hydration_fragment = current_hydration_fragment;
 
 	/** @type {Exports} */