+++ packages/svelte/src/internal/server/index.js
@@ -1,5 +1,4 @@
 import * as $ from '../client/runtime.js';
-import { set_is_ssr } from '../client/runtime.js';
 import { is_promise, noop } from '../common.js';
 import { subscribe_to_store } from '../../store/utils.js';
 import { DOMBooleanAttributes } from '../../constants.js';
@@ -80,6 +79,12 @@ export function assign_payload(p1, p2) {
 }
 
 /**
+ * Array of `onDestroy` callbacks that should be called at the end of the server render function
+ * @type {Function[]}
+ */
+export let on_destroy = [];
+
+/**
  * @param {(...args: any[]) => void} component
  * @param {{ props: Record<string, any>; context?: Map<any, any> }} options
  * @returns {RenderOutput}
@@ -89,7 +94,8 @@ export function render(component, options) {
 	const root_anchor = create_anchor(payload);
 	const root_head_anchor = create_anchor(payload.head);
 
-	set_is_ssr(true);
+	const prev_on_destroy = on_destroy;
+	on_destroy = [];
 	payload.out += root_anchor;
 
 	if (options.context) {
@@ -102,7 +108,8 @@ export function render(component, options) {
 		$.pop();
 	}
 	payload.out += root_anchor;
-	set_is_ssr(false);
+	for (const cleanup of on_destroy) cleanup();
+	on_destroy = prev_on_destroy;
 
 	return {
 		head: