+++ packages/svelte/tests/compiler-errors/samples/component-slot-duplicate-error-6/main.svelte
@@ -1,0 +1,9 @@
+<script>
+	import Nested from './irrelevant';
+	import Inner from './irrelevant';
+</script>
+
+<Nested>
+	<p slot="foo">{value}</p>
+	<Inner slot="foo">{value}</Inner>
+</Nested>