+++ packages/svelte/tests/compiler-errors/samples/runes-invalid-each-binding/main.svelte
@@ -1,0 +1,7 @@
+<script>
+	let arr = $state([1,2,3]);
+</script>
+
+{#each arr as value}
+	<input bind:value>
+{/each}