+++ packages/svelte/tests/hydration/samples/if-block-empty/main.svelte
@@ -1,0 +1,7 @@
+<script>
+	let foo = typeof window === 'undefined' ? '' : 'x';
+</script>
+
+{#if true}
+	{foo}
+{/if}