+++ packages/svelte/tests/runtime-legacy/samples/component-slot-let-named-2/SlotInner.svelte
@@ -1,0 +1,5 @@
+<script>
+	export let thing;
+</script>
+<span>{thing}</span>
+<slot />