+++ packages/svelte/tests/runtime-runes/samples/readonly-state-push/Child.svelte
@@ -1,0 +1,7 @@
+<script>
+	let { array } = $props();
+</script>
+
+{#each array as number}
+	<p>{number}</p>
+{/each}